tt_first.cpp
This is the first version of our project where the input is only for B Tech. Moreover, only ICT and MnC branches are being considered.
The program first takes input for the B Tech ICT and then displays TimeTable for them, then it takes input for B Tech and displays timetable for that, timetables of both are outputted in a .txt file timetable.txt.